// src/app/models/itemLink.ts
export class ItemLink{
    Name: string;
    Image: string;
    ImageName: string;
    RegionID: number;
    PowerBiLink: string;
    Order: number;
    OrderList: Array<number>;

//    constructor(name:string, flag: string, regionId:number, biLink:string){
//        this.Name = name;
//        this.LogoPath = flag;
//        this.RegionId = regionId;
//        this.PowerBiLink = biLink;
//    };
}